+++ src/warc2zim/content_rewriting/generic.py
@@ -64,6 +64,7 @@ class Rewriter:
         missing_zim_paths: set[ZimPath] | None,
         js_modules: set[ZimPath],
         charsets_to_try: list[str],
+        content_header_bytes_length: int,
         *,
         ignore_content_header_charsets: bool,
         ignore_http_header_charsets: bool,
@@ -83,6 +84,7 @@ class Rewriter:
         self.rewrite_mode = self.get_rewrite_mode(record, mimetype)
         self.js_modules = js_modules
         self.charsets_to_try = charsets_to_try
+        self.content_header_bytes_length = content_header_bytes_length
         self.ignore_content_header_charsets = ignore_content_header_charsets
         self.ignore_http_header_charsets = ignore_http_header_charsets
 
@@ -92,6 +94,7 @@ class Rewriter:
             self.content,
             self.encoding,
             self.charsets_to_try,
+            self.content_header_bytes_length,
             ignore_content_header_charsets=self.ignore_content_header_charsets,
             ignore_http_header_charsets=self.ignore_http_header_charsets,
         )