+++ tests/test_rewriting.py
@@ -39,6 +39,7 @@ def rewrite_generator():
             set(),
             set(),
             ["UTF-8", "ISO-8859-1"],
+            1024,
             ignore_http_header_charsets=False,
             ignore_content_header_charsets=False,
         ).rewrite(Template(""), Template(""))